requirements.txt
Flask
flask-login
bcrypt
nltk
python-dotenv
cryptography